# docs/openapi.yaml
openapi: 3.1.0
info:
  title: QuickServe API
  version: 1.0.0
  description: |
    QuickServe API - 地域の熟練ワーカーのためのサービス予約プラットフォーム

    ## 認証
    すべての保護されたエンドポイントは Bearer トークンを必要とします。
    `Authorization: Bearer <token>` ヘッダーを含めてください。
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.quickserve.com/api/v1
    description: Production server
...
paths:
...
  /workers/search:
    get:
      tags: [Workers]
      summary: 近隣ワーカー検索
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
        - name: latitude
          in: query
          required: true
          schema:
            type: number
        - name: longitude
          in: query
          required: true
          schema:
            type: number
        - name: radiusKm
          in: query
          schema:
            type: number
            default: 10
      responses:
        '200':
          description: ワーカー一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkerSearchResult'
  /workers/register:
    post:
      tags: [Workers]
      summary: ワーカー登録
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterWorkerInput'
      responses:
        '201':
          description: 登録成功
  /workers/{id}:
    get:
      tags: [Workers]
      summary: ワーカー詳細取得
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ワーカー詳細
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/WorkerProfile'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WorkerSearchResult:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
        averageRating:
          type: number
        totalReviews:
          type: integer
        distance:
          type: number
          description: Distance in kilometers
    WorkerProfile:
      type: object
      properties:
        id:
          type: string
        bio:
          type: string
        averageRating:
          type: number
        totalReviews:
          type: integer
        totalJobsCompleted:
          type: integer
        services:
          type: array
          items:
            $ref: '#/components/schemas/WorkerService'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
    WorkerService:
      type: object
      properties:
        id:
          type: string
        categoryId:
          type: string
        basePrice:
          type: number
        priceUnit:
          type: string
          enum: [per_job, per_hour]
        category:
          $ref: '#/components/schemas/Category'
    RegisterWorkerInput:
      type: object
      required: [idCardUrl, services]
      properties:
        bio:
          type: string
        idCardUrl:
          type: string
        services:
          type: array
          items:
            type: object
            required: [categoryId, basePrice]
            properties:
              categoryId:
                type: string
              basePrice:
                type: number
              priceUnit:
                type: string
                enum: [per_job, per_hour]
              description:
                type: string
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameLocal:
          type: string
        description:
          type: string
        iconUrl:
          type: string
        activeWorkersCount:
          type: integer
...
    Review:
      type: object
      properties:
        id:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        customer:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            avatarUrl:
              type: string